Extracted ReDownloading functionality into its own class
